securiter par controlle avec $_SESSION['fisrtAdmin']

+++ Controller/PostController.php
@@ -36,42 +36,50 @@ class PostController
     // AJOUT D'UN POST
     public function add()
     {   
-        if (!empty($_POST['title']) && !empty($_POST['content'])) {
-            $postId = $this->_postManager->add($_POST['title'], $_POST['content']);
-            if ($postId > 0) {
-                header('Location: index.php?objet=post&id=' . $postId);
-            }
+        if (isset($_SESSION['firstAdmin']) && $_SESSION['firstAdmin'] == 1 ) {
+            if (!empty($_POST['title']) && !empty($_POST['content'])) {
+                $postId = $this->_postManager->add($_POST['title'], $_POST['content']);
+                if ($postId > 0) {
+                    header('Location: index.php?objet=post&id=' . $postId);
+                }
 
-            throw new Exception('Impossible d\'ajouter le post !');
-        } 
-                
-        require_once('../view/formAddPost.php');
+                throw new Exception('Impossible d\'ajouter le post !');
+            } 
+            require_once('../view/formAddPost.php');
+        }
+        require_once('../view/noAccess.php');
     }
 
     // AFFICHAGE et MODIFICATION
     public function update($postId) 
-    {
-        if (!empty($_POST['title']) && !empty($_POST['content'])) {
-            $this->_postManager->update($postId, $_POST['title'], $_POST['content']);
+    {   
+        if (isset($_SESSION['firstAdmin']) && $_SESSION['firstAdmin'] == 1 ) {
+            if (!empty($_POST['title']) && !empty($_POST['content'])) {
+                $this->_postManager->update($postId, $_POST['title'], $_POST['content']);
 
-            header('Location: index.php?objet=post&id=' . $postId);
-        } 
-        $post = $this->_postManager->getPost($postId);
+                header('Location: index.php?objet=post&id=' . $postId);
+            } 
+            $post = $this->_postManager->getPost($postId);
 
-        require_once('../view/formUpdatePost.php');;
+            require_once('../view/formUpdatePost.php');
+        }
+        require_once('../view/noAccess.php');
     }
 
     // AFFICHAGE et SUPPRESSION
     public function delete($postId)
     {
-        if (!empty($_POST['title']) && !empty($_POST['content'])) {
-            $this->_postManager->delete($postId);
+        if (isset($_SESSION['firstAdmin']) && $_SESSION['firstAdmin'] == 1 ) {
+            if (!empty($_POST['title']) && !empty($_POST['content'])) {
+                $this->_postManager->delete($postId);
 
-            header('Location: index.php?objet=admin');
+                header('Location: index.php?objet=admin');
+            }
+            $post = $this->_postManager->getPost($postId);
+
+            require_once('../view/formDeletePost.php');
         }
-        $post = $this->_postManager->getPost($postId);
-
-        require_once('../view/formDeletePost.php');
+        require_once('../view/noAccess.php');
     }
 }
